pyproject.toml
[project]
name = "data-engineer-final-project-tourism-in-canary-islands"
version = "0.1.0"
description = "Data engineering pipeline for tourism data in the Canary Islands"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyspark==3.5.1",
    "google-cloud-storage>=2.16.0",
    "google-cloud-bigquery>=3.20.0",
    "dbt-bigquery==1.7.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
